+++ Modules/core/Controller/CoreconfigadminController.php
@@ -10,6 +10,7 @@ require_once 'Modules/core/Model/CoreInstall.php';
 
 require_once 'Modules/core/Model/CoreUser.php';
 require_once 'Modules/core/Model/CoreStatus.php';
+require_once 'Modules/core/Model/CoreConfig.php';
 
 /**
  *
@@ -266,20 +267,11 @@ class CoreconfigadminController extends CoresecureController {
 
         $value = $modelCoreConfig->getParam("user_desactivate", 1);
 
-        $choices = array();
-        $choicesid = array();
-        $choicesid[] = 1;
-        $choices[] = CoreTranslator::never($lang);
-        $choicesid[] = 2;
-        $choices[] = CoreTranslator::contract_ends($lang);
-        $choicesid[] = 3;
-        $choices[] = CoreTranslator::does_not_login_for_n_year(1, $lang);
-        $choicesid[] = 4;
-        $choices[] = CoreTranslator::does_not_login_for_n_year(2, $lang);
-        $choicesid[] = 5;
-        $choices[] = CoreTranslator::does_not_login_for_n_year(3, $lang);
-        $choicesid[] = 6;
-        $choices[] = CoreTranslator::contract_ends_or_does_not_login_for_1_year($lang);
+        $cc = new CoreConfig();
+        $expirationChoices = $cc->getExpirationChoices($lang);
+        $choices = $expirationChoices['labels'];
+        $choicesid = $expirationChoices['ids'];
+
 
         $form = new Form($this->request, "desactivateUserForm");
         $form->addSeparator(CoreTranslator::non_active_users($lang));